+++ src/neuroconv/datainterfaces/ecephys/spikeglx/spikeglxconverter.py
@@ -88,8 +88,11 @@ class SpikeGLXConverterPipe(ConverterPipe):
         data_interfaces = dict()
 
         nidq_streams = [stream_id for stream_id in streams_ids if stream_id == "nidq"]
-        electrical_streams = [stream_id for stream_id in streams_ids if stream_id not in nidq_streams]
-        for stream_id in electrical_streams:
+        sync_streams = [stream_id for stream_id in streams_ids if "SYNC" in stream_id]
+        neural_streams = [
+            stream_id for stream_id in streams_ids if stream_id not in nidq_streams and stream_id not in sync_streams
+        ]
+        for stream_id in neural_streams:
             data_interfaces[stream_id] = SpikeGLXRecordingInterface(folder_path=folder_path, stream_id=stream_id)
 
         for stream_id in nidq_streams: